Update README.md
This is what I call the 'intermediate cardinal' write-up for the narrative, what I would verbally tell someone from Thyssen-Krupp. If I proceed with this branch, the next step will entail fleshing out the actual routines--writing the disparate elements of the narrative more suitably for a Finite State Machine diagram.