Update README.md
This is the 'intermediate differential' version of the narrative. The control narrative itself has two states--'idle' and 'moving'. 'Moving' has two tables, one with the floors to stop at *before* the drive changes direction (the 'current' table), and another with floors to move to the 'current' table, which the elevator will stop at, *after* the drive changes direction. In this case, the control narrative does not concern whether the elevator is moving up or down--only stopping at the floors listed on the current table, and what a new entry will do to the rate-of-change between the current or previous floor and the next floor--if the rate-of-change would become less than or equal to zero, the entry is added to the 'next' table, not the 'current' table, and the floor will be stopped at after the drive changes direction.